+++ packages/veritone-widgets/src/build-entry.js
@@ -24,3 +24,10 @@ export EngineOutputExport, {
 export engineOutputExportReducer, * as engineOutputExportModule from './redux/modules/engineOutputExport';
 export engineOutputExportSaga from './redux/modules/engineOutputExport/saga';
 export { UserProfileWidget } from './widgets/UserProfile';
+export multipleEngineSelectionReducer, * as multipleEngineSelectionModule from './redux/modules/multipleEngineSelection';
+export EnginePicker, {
+  EnginePickerWidget,
+  SelectionInfoPanel,
+  SelectionInfoPanelWidget
+} from './widgets/EnginePicker';
+export { InfinitePicker } from './widgets/EnginePicker/InfinitePicker';